[TASK] Drop obsolete images
The images `php53` and `php54` have died with the EOL of TYPO3 6.2 ELTS.
These images are not used anymore in CI and can be removed.
As Bamboo was shut down earlier and thus no agents are available anymore,
the `bamboo-remote-agent` is obsolete as well.

This renders `baseimage` obsolete, as the remaining `php` images derive
from `php:x.x-cli` now.